# Add StaticVisitor for faster validation