GraphQL::Language::Visitor has support for modifying the document during visitation. But for Validation and analysis, this isn't necessary -- it never modifies the document. 

So, this adds a lighter-weight visitor which doesn't include support for modifications, but runs faster and with less memory: 

```ruby
require "bundler/inline"

gemfile do
  gem "graphql", path: "./"
  gem "benchmark-ips"
end

doc = GraphQL.parse(GraphQL::Introspection::INTROSPECTION_QUERY)

Benchmark.ips do |x|
  x.report("Old Visitor") { GraphQL::Language::Visitor.new(doc).visit }
  x.report("New Visitor") { GraphQL::Language::StaticVisitor.new(doc).visit }
  x.compare!
end
```

```
Warming up --------------------------------------
         Old Visitor     1.791k i/100ms
         New Visitor     5.342k i/100ms
Calculating -------------------------------------
         Old Visitor     17.760k (± 4.0%) i/s -     89.550k in   5.051364s
         New Visitor     54.783k (± 6.0%) i/s -    277.784k in   5.091691s

Comparison:
         New Visitor:    54783.2 i/s
         Old Visitor:    17760.0 i/s - 3.08x  slower
```


So, the new visitor is 3x faster. But, I couldn't see notable runtime improvements in existing benchmarks:

```
$ be rake bench:profile_small_result
```

|   | `master` | this branch |
|---|---|---|
| YJIT | 867.621  (± 3.2%) i/s | 875.228  (± 4.8%) i/s |
| no | 398.409  (± 4.8%) i/s  | 398.074  (± 3.3%) i/s | 

```diff 
- Total allocated: 109952 bytes (1131 objects)
+ Total allocated: 109080 bytes (1111 objects)
```

```
$ be rake bench:profile_small_introspection
```

|   | `master` | this branch |
|---|---|---|
| YJIT | 218.893  (± 5.0%) i/s | 219.840  (± 3.6%) i/s |
| no | 95.241  (± 3.1%) i/s  |  97.275  (± 4.1%) i/s | 

```diff 
- Total allocated: 295456 bytes (2827 objects)
+ Total allocated: 292144 bytes (2749 objects)
```

But the reduced memory consumption is worth it, and I bet there _are_ cases where the performance improvement would be noticeable (where the document is very large but the returned data is very small).